+++ src/domain/dtos/AuthConfirmPhoneDto.ts
@@ -1,0 +1,6 @@
+import {PhoneField} from '@steroidsjs/nest/infrastructure/decorators/fields';
+
+export class AuthConfirmPhoneDto {
+    @PhoneField()
+    phone: string;
+}