+++ src/domain/enums/AuthConfirmProviderTypeEnum.ts
@@ -1,0 +1,6 @@
+export enum AuthConfirmProviderTypeEnum {
+    CALL ='call',
+    SMS = 'sms',
+    VOICE = 'voice',
+    EMAIL = 'email'
+}