+++ src/domain/types/AuthConfirmProviderType.ts
@@ -1,0 +1,3 @@
+import {AuthConfirmProviderTypeEnum} from '../enums/AuthConfirmProviderTypeEnum';
+
+export type AuthConfirmProviderType = AuthConfirmProviderTypeEnum | string;