+++ src/infrastructure/controllers/AuthPhoneController.ts
@@ -1,9 +1,10 @@
 import {ApiBody, ApiOkResponse, ApiTags} from '@nestjs/swagger';
-import {Body, Controller, Inject, Post, UseGuards} from '@nestjs/common';
 import NotifierProviderType from '@steroidsjs/nest-modules/notifier/enums/NotifierProviderType';
+import {Body, Controller, Inject, Post, Res, UseGuards} from '@nestjs/common';
 import {ContextDto} from '@steroidsjs/nest/usecases/dtos/ContextDto';
 import {Context} from '@steroidsjs/nest/infrastructure/decorators/Context';
 import {DataMapper} from '@steroidsjs/nest/usecases/helpers/DataMapper';
+import {Response} from 'express';
 import {CodeAuthGuard} from '../guards/CodeAuthGuard';
 import {AuthenticateWithCodeDto} from '../../usecases/sendAuthenticationCodeUseCase/dtos/AuthenticateWithCodeDto';
 import {AuthConfirmLoginDto} from '../../domain/dtos/AuthConfirmLoginDto';
@@ -17,6 +18,9 @@ import {
     ISendAuthenticationCodeUseCase,
     SEND_AUTHENTICATION_CODE_USE_CASE_TOKEN,
 } from '../../usecases/sendAuthenticationCodeUseCase/ISendAuthenticationCodeUseCase';
+import {AuthConfirmPhoneDto} from '../../domain/dtos/AuthConfirmPhoneDto';
+import {AuthCookieLoginSchema} from '../schemas/AuthCookieLoginSchema';
+import {AuthCookieService} from '../services/AuthCookieService';
 
 @ApiTags('Авторизация по телефону')
 @Controller('/auth/phone')
@@ -26,18 +30,19 @@ export class AuthPhoneController {
         private readonly authenticateWithCodeUseCase: IAuthenticateWithCodeUseCase,
         @Inject(SEND_AUTHENTICATION_CODE_USE_CASE_TOKEN)
         private readonly sendAuthenticationCodeUseCase: ISendAuthenticationCodeUseCase,
+        private readonly authCookieService: AuthCookieService,
     ) {
     }
 
     @Post('/sms')
     @ApiOkResponse({type: AuthConfirmSchema})
     async sendSmsCode(
-        @Body() dto: AuthenticateWithCodeDto,
+        @Body() dto: AuthConfirmPhoneDto,
         @Context() context: ContextDto,
     ) {
         const authConfirm = await this.sendAuthenticationCodeUseCase.handle(
             NotifierProviderType.SMS,
-            dto,
+            DataMapper.create(AuthenticateWithCodeDto, {target: dto.phone}),
             context,
         );
         return DataMapper.create(AuthConfirmSchema, authConfirm);
@@ -46,12 +51,12 @@ export class AuthPhoneController {
     @Post('/call')
     @ApiOkResponse({type: AuthConfirmSchema})
     async sendSmsCodeByCall(
-        @Body() dto: AuthenticateWithCodeDto,
+        @Body() dto: AuthConfirmPhoneDto,
         @Context() context: ContextDto,
     ) {
         const authConfirm = await this.sendAuthenticationCodeUseCase.handle(
             NotifierProviderType.CALL,
-            dto,
+            DataMapper.create(AuthenticateWithCodeDto, {target: dto.phone}),
             context,
         );
         return DataMapper.create(AuthConfirmSchema, authConfirm);
@@ -60,12 +65,12 @@ export class AuthPhoneController {
     @Post('/send')
     @ApiOkResponse({type: AuthConfirmSchema})
     async send(
-        @Body() dto: AuthenticateWithCodeDto,
+        @Body() dto: AuthConfirmPhoneDto,
         @Context() context: ContextDto,
     ) {
         const authConfirm = await this.sendAuthenticationCodeUseCase.handle(
             null,
-            dto,
+            DataMapper.create(AuthenticateWithCodeDto, {target: dto.phone}),
             context,
         );
         return DataMapper.create(AuthConfirmSchema, authConfirm);
@@ -81,5 +86,20 @@ export class AuthPhoneController {
     ) {
         const authLogin = await this.authenticateWithCodeUseCase.handle(dto, context);
         return DataMapper.create(AuthLoginSchema, authLogin);
+    }
+
+    @Post('/confirm/cookie')
+    @ApiBody({type: AuthConfirmLoginDto})
+    @ApiOkResponse({type: AuthCookieLoginSchema})
+    @UseGuards(CodeAuthGuard)
+    async authenticateWithCodeCookie(
+        @Body() dto: AuthConfirmLoginDto,
+        @Context() context: ContextDto,
+        @Res({passthrough: true}) response: Response,
+    ) {
+        const authLogin = await this.authenticateWithCodeUseCase.handle(dto, context);
+        this.authCookieService.setRefreshToken(response, authLogin.refreshToken);
+        this.authCookieService.setAccessToken(response, authLogin.accessToken);
+        return DataMapper.create(AuthCookieLoginSchema, authLogin);
     }
 }