+++ src/infrastructure/module.ts
@@ -1,16 +1,12 @@
 import {JwtModule} from '@nestjs/jwt';
 import {PassportModule} from '@nestjs/passport';
-import {ModuleHelper} from '@steroidsjs/nest/infrastructure/helpers/ModuleHelper';
 import {UserModule} from '@steroidsjs/nest-modules/user/UserModule';
 import {FileModule} from '@steroidsjs/nest-modules/file/FileModule';
 import {NotifierModule} from '@steroidsjs/nest-modules/notifier/NotifierModule';
-import {IUserService} from '@steroidsjs/nest-modules/user/services/IUserService';
-import {IFileService} from '@steroidsjs/nest-modules/file/services/IFileService';
-import {INotifierService} from '@steroidsjs/nest-modules/notifier/services/INotifierService';
-import {forwardRef} from '@nestjs/common';
+import {forwardRef, ModuleMetadata} from '@nestjs/common';
 import {IAuthUpdateUserOwnPasswordUseCase} from '@steroidsjs/nest-modules/auth/usecases/IAuthUpdateUserOwnPasswordUseCase';
-import {IUserUpdatePasswordUseCase} from '@steroidsjs/nest-modules/user/usecases/IUserUpdatePasswordUseCase';
 import {IAuthRevokeUserActiveLoginsUseCase} from '@steroidsjs/nest-modules/auth/usecases/IAuthRevokeUserActiveLoginsUseCase';
+import {IValidator} from '@steroidsjs/nest/usecases/interfaces/IValidator';
 import {AuthService} from '../domain/services/AuthService';
 import {AuthLoginService} from '../domain/services/AuthLoginService';
 import {AuthPermissionsService} from '../domain/services/AuthPermissionsService';
@@ -22,12 +18,19 @@ import {AuthConfirmService} from '../domain/services/AuthConfirmService';
 import {IAuthRoleRepository} from '../domain/interfaces/IAuthRoleRepository';
 import {AuthRoleService} from '../domain/services/AuthRoleService';
 import {AuthFilePermissionService} from '../domain/services/AuthFilePermissionService';
-import {AuthUpdateUserOwnPasswordUseCase} from '../usecases/updatePassword/AuthUpdateUserOwnPasswordUseCase';
-import {AuthRevokeUserActiveLoginsUseCase} from '../usecases/revokeUserActiveLogins/AuthRevokeUserActiveLoginsUseCase';
 import {AuthenticateWithCodeUseCase} from '../usecases/authenticateWithCodeUseCase/AuthenticateWithCodeUseCase';
 import {SendAuthenticationCodeUseCase} from '../usecases/sendAuthenticationCodeUseCase/SendAuthenticationCodeUseCase';
 import {AUTHENTICATE_WITH_CODE_USE_CASE_TOKEN} from '../usecases/authenticateWithCodeUseCase/IAuthenticateWithCodeUseCase';
 import {SEND_AUTHENTICATION_CODE_USE_CASE_TOKEN} from '../usecases/sendAuthenticationCodeUseCase/ISendAuthenticationCodeUseCase';
+import {AuthUpdateUserOwnPasswordUseCase} from '../usecases/updatePassword/AuthUpdateUserOwnPasswordUseCase';
+import {AuthRevokeUserActiveLoginsUseCase} from '../usecases/revokeUserActiveLogins/AuthRevokeUserActiveLoginsUseCase';
+import {AUTH_CONFIRM_PROVIDERS_TOKEN, IAuthConfirmProvider} from '../domain/interfaces/IAuthConfirmProvider';
+import {
+    GET_AUTH_CONFIRM_TARGET_FIELD_USE_CASE_TOKEN,
+} from '../usecases/getAuthConfirmTargetField/IGetAuthConfirmTargetFieldUseCase';
+import {
+    GetAuthConfirmTargetFieldUseCase,
+} from '../usecases/getAuthConfirmTargetField/GetAuthConfirmTargetFieldUseCase';
 import {SessionService} from './services/SessionService';
 import {AuthLoginRepository} from './repositories/AuthLoginRepository';
 import {AuthPermissionRepository} from './repositories/AuthPermissionRepository';
@@ -36,15 +39,22 @@ import {JwtStrategy} from './strategies/JwtStrategy';
 import {AuthConfirmRepository} from './repositories/AuthConfirmRepository';
 import {LoginSmsCodeStrategy} from './strategies/LoginSmsCodeStrategy';
 import {AuthRoleRepository} from './repositories/AuthRoleRepository';
+import {authConfirmProviders} from './services/authConfirmProviders';
 import {AuthController} from './controllers/AuthController';
+import {AuthEmailController} from './controllers/AuthEmailController';
 import {AuthFilePermissionController} from './controllers/AuthFilePermissionController';
 import {AuthPermissionController} from './controllers/AuthPermissionController';
 import {AuthPhoneController} from './controllers/AuthPhoneController';
 import {AuthRoleController} from './controllers/AuthRoleController';
 import {IAuthModuleConfig} from './config';
-import {PasswordValidator} from './validators/PasswordValidator';
+import {authUpdatePasswordValidators} from './validators';
+import {AUTH_UPDATE_PASSWORD_VALIDATORS_TOKEN} from '../domain/constants/AuthUpdatePasswordValidatorsToken';
+import {GeneratePermissionsMigrationCommand} from './commands/GeneratePermissionsMigrationCommand';
+import {AuthNewPermissionsCheckService} from './services/AuthNewPermissionsCheckService';
+import {AuthCookieController} from './controllers/AuthCookieController';
+import {AuthCookieService} from './services/AuthCookieService';
 
-export default (config: IAuthModuleConfig) => ({
+export default (config: IAuthModuleConfig): ModuleMetadata => ({
     imports: [
         PassportModule,
         NotifierModule,
@@ -56,6 +66,8 @@ export default (config: IAuthModuleConfig) => ({
     ],
     controllers: [
         AuthController,
+        AuthEmailController,
+        AuthCookieController,
         AuthFilePermissionController,
         AuthPermissionController,
         AuthPhoneController,
@@ -82,55 +94,36 @@ export default (config: IAuthModuleConfig) => ({
             provide: IAuthConfirmRepository,
             useClass: AuthConfirmRepository,
         },
-        ModuleHelper.provide(AuthRoleService, [
-            IAuthRoleRepository,
-            AuthPermissionsService,
-        ]),
-        ModuleHelper.provide(AuthService, [
-            IUserService,
-            ISessionService,
-            AuthLoginService,
-            AuthPermissionsService,
-        ]),
-        ModuleHelper.provide(AuthConfirmService, [
-            IAuthConfirmRepository,
-            INotifierService,
-            AuthService,
-        ]),
-        ModuleHelper.provide(AuthLoginService, [
-            IAuthLoginRepository,
-            ISessionService,
-        ]),
-        ModuleHelper.provide(AuthPermissionsService, [
-            IAuthPermissionsRepository,
-            IAuthRoleRepository,
-        ]),
-        ModuleHelper.provide(AuthFilePermissionService, [
-            IFileService,
-        ]),
-        ModuleHelper.provide(LoginPasswordStrategy, [
-            IUserService,
-            AuthService,
-            ISessionService,
-        ]),
-        ModuleHelper.provide(LoginSmsCodeStrategy, [
-            AuthConfirmService,
-            AuthService,
-            ISessionService,
-        ]),
+
+        AuthRoleService,
+        AuthService,
+        ...authConfirmProviders,
+        {
+            provide: AUTH_CONFIRM_PROVIDERS_TOKEN,
+            useFactory: (...providers: IAuthConfirmProvider[]) => providers,
+            inject: authConfirmProviders,
+        },
+        {
+            provide: GET_AUTH_CONFIRM_TARGET_FIELD_USE_CASE_TOKEN,
+            useClass: GetAuthConfirmTargetFieldUseCase,
+        },
+        AuthConfirmService,
+        AuthLoginService,
+        AuthPermissionsService,
+        AuthFilePermissionService,
+        LoginPasswordStrategy,
+        LoginSmsCodeStrategy,
         JwtStrategy,
 
         // UseCases
-        ModuleHelper.provide(AuthUpdateUserOwnPasswordUseCase, IAuthUpdateUserOwnPasswordUseCase, [
-            IUserUpdatePasswordUseCase,
-            IAuthRevokeUserActiveLoginsUseCase,
-            [
-                PasswordValidator,
-            ],
-        ]),
-        ModuleHelper.provide(AuthRevokeUserActiveLoginsUseCase, IAuthRevokeUserActiveLoginsUseCase, [
-            AuthLoginService,
-        ]),
+        {
+            provide: IAuthUpdateUserOwnPasswordUseCase,
+            useClass: AuthUpdateUserOwnPasswordUseCase,
+        },
+        {
+            provide: IAuthRevokeUserActiveLoginsUseCase,
+            useClass: AuthRevokeUserActiveLoginsUseCase,
+        },
         {
             provide: AUTHENTICATE_WITH_CODE_USE_CASE_TOKEN,
             useClass: AuthenticateWithCodeUseCase,
@@ -141,10 +134,15 @@ export default (config: IAuthModuleConfig) => ({
         },
 
         // Validators
-        ModuleHelper.provide(PasswordValidator, [
-            IUserService,
-            ISessionService,
-        ]),
+        ...authUpdatePasswordValidators,
+        {
+            provide: AUTH_UPDATE_PASSWORD_VALIDATORS_TOKEN,
+            useFactory: (...providers: IValidator[]) => providers,
+            inject: authUpdatePasswordValidators,
+        },
+        GeneratePermissionsMigrationCommand,
+        AuthNewPermissionsCheckService,
+        AuthCookieService,
     ],
     exports: [
         ISessionService,