+++ src/usecases/updatePassword/AuthUpdateUserOwnPasswordUseCase.ts
@@ -4,12 +4,18 @@ import {IUserUpdatePasswordUseCase} from '@steroidsjs/nest-modules/user/usecases
 import {IAuthUpdateUserOwnPasswordUseCase} from '@steroidsjs/nest-modules/auth/usecases/IAuthUpdateUserOwnPasswordUseCase';
 import {IAuthRevokeUserActiveLoginsUseCase} from '@steroidsjs/nest-modules/auth/usecases/IAuthRevokeUserActiveLoginsUseCase';
 import {ContextDto} from '@steroidsjs/nest/usecases/dtos/ContextDto';
+import {Inject, Injectable} from '@nestjs/common';
 import {AuthUpdateUserOwnPasswordUseCaseDto} from './dtos/AuthUpdateUserOwnPasswordUseCaseDto';
+import {AUTH_UPDATE_PASSWORD_VALIDATORS_TOKEN} from '../../domain/constants/AuthUpdatePasswordValidatorsToken';
 
+@Injectable()
 export class AuthUpdateUserOwnPasswordUseCase implements IAuthUpdateUserOwnPasswordUseCase {
     constructor(
+        @Inject(IUserUpdatePasswordUseCase)
         private readonly userUpdatePasswordUseCase: IUserUpdatePasswordUseCase,
+        @Inject(IAuthRevokeUserActiveLoginsUseCase)
         private readonly revokeUserActiveLoginsUseCase: IAuthRevokeUserActiveLoginsUseCase,
+        @Inject(AUTH_UPDATE_PASSWORD_VALIDATORS_TOKEN)
         private readonly validators: IValidator[],
     ) {}